LayoutTests/imported/w3c/web-platform-tests/streams/readable-streams/async-iterator.any.sharedworker-expected.txt


PASS Async iterator instances should have the correct list of properties
PASS Async-iterating a push source
PASS Async-iterating a pull source
FAIL Async-iterating a push source with undefined values assert_array_equals: lengths differ, expected array [undefined, undefined, undefined] length 3, got [] length 0
FAIL Async-iterating a pull source with undefined values assert_array_equals: lengths differ, expected array [undefined, undefined, undefined] length 3, got [] length 0
PASS Async-iterating a pull source manually
FAIL Async-iterating an errored stream throws assert_equals: expected (string) "e" but got (object) object "TypeError: Type error"
PASS Async-iterating a closed stream never executes the loop body, but works fine
PASS Async-iterating an empty but not closed/errored stream never executes the loop body and stalls the async function
PASS Async-iterating a partially consumed stream
PASS Cancellation behavior when throwing inside loop body; preventCancel = false
FAIL Cancellation behavior when throwing inside loop body; preventCancel = true assert_array_equals: cancel() should not be called lengths differ, expected array ["pull"] length 1, got ["pull", "cancel", undefined] length 3
PASS Cancellation behavior when breaking inside loop body; preventCancel = false
FAIL Cancellation behavior when breaking inside loop body; preventCancel = true assert_array_equals: cancel() should not be called lengths differ, expected array ["pull"] length 1, got ["pull", "cancel", undefined] length 3
PASS Cancellation behavior when returning inside loop body; preventCancel = false
FAIL Cancellation behavior when returning inside loop body; preventCancel = true assert_array_equals: cancel() should not be called lengths differ, expected array ["pull"] length 1, got ["pull", "cancel", undefined] length 3
PASS Cancellation behavior when manually calling return(); preventCancel = false
FAIL Cancellation behavior when manually calling return(); preventCancel = true assert_array_equals: cancel() should not be called lengths differ, expected array [] length 0, got ["cancel", undefined] length 2
FAIL next() rejects if the stream errors promise_rejects_exactly: 2nd next() function "function() { throw e; }" threw object "TypeError: Type error" but we expected it to throw object "Error: error1"
FAIL return() does not rejects if the stream has not errored yet assert_equals: done expected true but got false
PASS return() rejects if the stream has errored
FAIL next() that succeeds; next() that reports an error; next() promise_rejects_exactly: 2nd next() function "function() { throw e; }" threw object "TypeError: Type error" but we expected it to throw object "Error: error1"
FAIL next() that succeeds; next() that reports an error(); next() [no awaiting] assert_equals: 2nd next() rejection reason expected object "Error: error1" but got object "TypeError: Type error"
FAIL next() that succeeds; next() that reports an error(); return() promise_rejects_exactly: 2nd next() function "function() { throw e; }" threw object "TypeError: Type error" but we expected it to throw object "Error: error1"
FAIL next() that succeeds; next() that reports an error(); return() [no awaiting] assert_equals: 2nd next() rejection reason expected object "Error: error1" but got object "TypeError: Type error"
FAIL next() that succeeds; return() assert_equals: return() done expected true but got false
FAIL next() that succeeds; return() [no awaiting] assert_equals: return() done expected true but got false
FAIL return(); next() assert_equals: return() done expected true but got false
FAIL return(); next() [no awaiting] assert_equals: return() done expected true but got false
FAIL return(); next() with delayed cancel() assert_equals: return() done expected true but got false
FAIL return(); next() with delayed cancel() [no awaiting] assert_equals: return() should resolve with original reason done expected true but got false
FAIL return(); return() assert_equals: 1st return() done expected true but got false
FAIL return(); return() [no awaiting] assert_equals: 1st return() done expected true but got false
PASS values() throws if there's already a lock
PASS Acquiring a reader after exhaustively async-iterating a stream
FAIL Acquiring a reader after return()ing from a stream that errors promise_rejects_exactly: 2nd next() function "function() { throw e; }" threw object "TypeError: Type error" but we expected it to throw object "Error: error1"
PASS Acquiring a reader after partially async-iterating a stream
FAIL Acquiring a reader and reading the remaining chunks after partially async-iterating a stream with preventCancel = true assert_equals: value expected (number) 3 but got (undefined) undefined
PASS return() should unlock the stream synchronously when preventCancel = false
PASS return() should unlock the stream synchronously when preventCancel = true
PASS close() while next() is pending
